+++ metrics/ttfp/run-benchmark.jl
@@ -214,7 +214,8 @@ ENV["JULIA_PKG_PRECOMPILE_AUTO"] = 0
 project1 = make_project_folder("current-pr")
 Pkg.activate(project1)
 if Package == "WGLMakie"
-    Pkg.add([(; name="Electron"), (; name="JSServe", rev="master")])
+    Pkg.add([(; name="Electron"),
+             (; url="https://github.com/SimonDanisch/Bonito.jl", name="Bonito", rev="master")])
 end
 pkgs = NamedTuple[(; path="./MakieCore"), (; path="."), (; path="./$Package")]
 # cd("dev/Makie")